+++ experiments/benchmarks/bucket.jl
@@ -46,7 +46,7 @@ earth_param_set = ClimaLand.Parameters.LandParameters(FT);
 outdir = "bucket_benchmark_$(device_suffix)"
 !ispath(outdir) && mkpath(outdir)
 
-function setup_prob(t0, tf, Δt; nelements = (101, 7))
+function setup_prob(t0, tf, Δt; nelements = (200, 7))
     # We set up the problem in a function so that we can make multiple copies (for profiling)
 
     # Set up simulation domain
@@ -147,7 +147,7 @@ function setup_simulation(; greet = false)
     t0 = 0.0
     tf = 7 * 86400
     Δt = 3600.0
-    nelements = (101, 7)
+    nelements = (200, 7)
     if greet
         @info "Run: Bucket with temporal map"
         @info "Resolution: $nelements"
@@ -244,7 +244,7 @@ if ClimaComms.device() isa ClimaComms.CUDADevice
 end
 
 if get(ENV, "BUILDKITE_PIPELINE_SLUG", nothing) == "climaland-benchmark"
-    PREVIOUS_BEST_TIME = 0.477
+    PREVIOUS_BEST_TIME = 1.69
     if average_timing_s > PREVIOUS_BEST_TIME + std_timing_s
         @info "Possible performance regression, previous average time was $(PREVIOUS_BEST_TIME)"
     elseif average_timing_s < PREVIOUS_BEST_TIME - std_timing_s